Refactored creation dialogs into seperate module
