+++ src/app/home/home.component.html
@@ -9,7 +9,7 @@
       [opened]="!handset.matches"
     >
       <mat-toolbar color="primary" class="toolbar">
-        <a href="/"><img class="logo" src="assets/logos/white.svg" alt="logo" /></a>
+        <a routerLink="/"><img alt="logo" class="logo" src="assets/logos/white.svg"/></a>
       </mat-toolbar>
       <div class="new-button-container">
         <app-create-entity-menu (addLabelClicked)="addLabelClicked()" (addEpicClicked)="addEpicClicked()"></app-create-entity-menu>
