+++ src/app/shared/creation-dialogs/components/epic-dialog/epic-dialog.component.ts
@@ -1,8 +1,8 @@
-import { ChangeDetectionStrategy, Component, Inject, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { EpicDialogData } from '../../models';
+import { EpicDialogData } from '../../../../pages/epic/models';
 import { Epic } from '@state';
-import { EpicDialogResult } from '../../models/epic-dialog';
+import { EpicDialogResult } from '../../../../pages/epic/models/epic-dialog';
 
 @Component({
   selector: 'app-epic-dialog',
@@ -12,6 +12,7 @@ import { EpicDialogResult } from '../../models/epic-dialog';
 })
 export class EpicDialogComponent {
   epic: Epic;
+
   constructor(public dialogRef: MatDialogRef<EpicDialogComponent, EpicDialogResult>, @Inject(MAT_DIALOG_DATA) public data: EpicDialogData) {
     if (data.case === 'add') {
       this.epic = {
